requirements.txt
# YOLOv8 Solar Panel Detection - Requirements
# Python 3.10 or 3.11 recommended

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# COCO Format Support
pycocotools>=2.0.6

# Progress Bars
tqdm>=4.65.0

# YAML Configuration
PyYAML>=6.0

# Image Augmentation (optional but recommended)
albumentations>=1.3.0

# Jupyter Support (optional - for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0

# Additional utilities
scipy>=1.11.0
